+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2
 
 Change description package.
+
+## 0.0.3
+* Fix startFromrRight angle icon
+* Change icon property from widget to iconData
+* Add property iconColor, backgroundIconColor, borderRadiusBackgroundIcon